requirements.txt
sqlalchemy
Click==7.0
dataclasses==0.6
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-Login==0.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
